# Parallelize local base image compression