--- a/jib-core/src/main/java/com/google/cloud/tools/jib/builder/steps/StepsRunner.java
+++ b/jib-core/src/main/java/com/google/cloud/tools/jib/builder/steps/StepsRunner.java
@@ -257,6 +257,7 @@ public class StepsRunner {
         executorService.submit(
             () ->
                 new ExtractTarStep(
+                        executorService,
                         buildConfiguration,
                         results.tarPath.get(),
                         childProgressDispatcherFactory,