Towards #1913. This only really makes a difference for un-cached local base image layers, but it provides a decent speedup in that case. Here were the build times using maven + a clean cache on the master branch vs. these changes:

`docker://gcr.io/distroless/java` 
* master: **7.6 seconds**
* this PR: **6.8 seconds**

`docker://openjdk:8`
* master: **29.2 seconds**
* this PR: **16.4 seconds**